Refactor: Move `formatPaymentsTextForTicket` to `DailyReportScreen`
The `formatPaymentsTextForTicket` function was moved from `ThermalPrinting` object to `DailyReportScreen` to improve code organization and reduce coupling.

This change also involved:
- Adding `@RequiresApi(Build.VERSION_CODES.S)` annotation to `AppNavigation` and `DailyReportScreen` composables.
- Updating calls to `formatIsoDate` to use `DateUtils.formatIsoDate` for clarity.
- Making `ThermalPrinting.centerText` public.
